Finished deletes in groupchat. Need to tidy up code and comment routes

+++ daos/groupChat.py
@@ -80,27 +80,34 @@ class groupChatDAO:
 
     def deleteGroupChat(self, gc_id):
         cursor = self.conn.cursor()
-        query = "delete from groupchat where gc_id = %s;"
+        query = "delete from groupchat where gc_id = %s returning gc_id;"
         cursor.execute(query, (gc_id,))
         self.conn.commit()
 
+        gid = cursor.fetchone()[0]
+        return gid
+
     def addUserToGroupChat(self, gc_id, user_id):
         cursor = self.conn.cursor()
-        query = "insert into belongsto(gc_id, user_id) values(%s,%s);"
+        query = "insert into belongsto(gc_id, user_id) values(%s,%s) returning user_id;"
+
         cursor.execute(query, (gc_id, user_id,))
+        user = cursor.fetchone()[0]
         self.conn.commit()
+        return user
 
     def deleteUserFromGroupChat(self, gc_id, user_id):
         cursor = self.conn.cursor()
-        query = "delete from belongsto where gc_id=%s AND user_id=%s;"
+        query = "delete from belongsto where gc_id=%s AND user_id=%s returning user_id;"
         cursor.execute(query, (gc_id, user_id,))
         self.conn.commit()
+        id = cursor.fetchone()[0]
+        return id
 
     def deleteAllUsersFromGroupChat(self, gc_id):
         cursor = self.conn.cursor()
         query ="delete from belongsto where gc_id=%s;"
         cursor.execute(query, (gc_id,))
-        gid = cursor.fetchone()[0]
         self.conn.commit()
 
     def getGroupChatsForUser(self, user_id):
